further refactor and move test to the right file
